pyproject.toml
[tool.poetry]
name = "oracle"
version = "2.0.0"
description = "StakeWise Oracles are responsible for submitting off-chain data."
authors = ["Dmitri Tsumak <tsumak.dmitri@gmail.com>"]
license = "AGPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.8.12"
python-decouple = "==3.5"
web3 = "==5.25.0"
py-ecc = "==5.2.0"
gql = {version = "3.0.0b1", allow-prereleases = true, extras = ["aiohttp"]}
backoff = "==1.11.1"
aiohttp = {version = "==3.7.4.post0", extras = ["speedups"]}
boto3 = {version = "==1.20.20", extras = ["awscrt"]}

[tool.poetry.dev-dependencies]
mypy = "==0.910"
black = "==21.11b1"
flake8 = "==4.0.1"
isort = "==5.10.1"
flake8-black = "==0.2.3"
flake8-bugbear = "==21.11.29"
pre-commit = "==2.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
